requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
openai-whisper==20231117
torch==2.2.0
ffmpeg-python==0.2.0
pydantic==2.6.3
aiofiles==23.2.1
lxml==5.1.0
